backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Backend API for RAG Chatbot for Published Book"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "google-generativeai>=0.5.0",
    "qdrant-client>=1.8.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.3.0",
    "asyncpg>=0.29.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]